+++ .eslintrc.js
@@ -19,7 +19,8 @@ module.exports = {
     webextensions: true,
     browser: true,
     es6: true,
-    node: true
+    node: true,
+    jest: true
   },
   rules: {
     "prefer-const": error
