+++ package.json
@@ -11,7 +11,7 @@
   "main": "dist/bundle.js",
   "scripts": {
     "build": "webpack --config webpack.config.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -41,6 +41,8 @@
     "file-loader": "^4.2.0",
     "fluture": "^11.0.3",
     "html-webpack-plugin": "^3.2.0",
+    "jest": "^24.9.0",
+    "node-fetch": "^2.6.0",
     "node-sass": "^4.12.0",
     "prettier": "^1.18.2",
     "prop-types": "^15.7.2",
