+++ src/components/variables.js
@@ -1,0 +1,3 @@
+const RATES = ["BTC", "BCH", "DAI", "ETC", "ETH", "LTC", "XRP", "XLM"];
+
+export default RATES;
